[Refactor] ColorModal 파일명 변경
